Move LFPManager to a new file read_lfp_factors only once and filter per section
